update jobs response
